+++ CHANGELOG.md
@@ -15,7 +15,7 @@
 - [](https://github.com/vegaprotocol/vegawallet/pull/) -
 
 ### 🐛 Fixes
-- [](https://github.com/vegaprotocol/vegawallet/pull/) -
+- [558](https://github.com/vegaprotocol/vegawallet/pull/558) - Use user-friendly RPC code as error message
 
 ## 0.15.0